Updates to excel export - styling, dates, moving code into utils file
